requirements.txt
alembic==1.11.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.2
orjson==3.8.5
packaging==23.1
pandas==1.5.3
pluggy==1.0.0
psycopg==3.1.9
pydantic==1.10.4
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette
tinydb==4.7.1
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2023.3
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
win32-setctime==1.1.0
xlrd==2.0.1
